refactor to resolve maven issue
