+++ src/main/java/stepDefinitions/WebDriver/World.java
@@ -1,0 +1,7 @@
+package stepDefinitions.WebDriver;
+
+import java.util.HashMap;
+
+public class World {
+    public HashMap<String,Object> context = new HashMap<String, Object>();
+}
