--- src/test/java/stepDefinitions/WebDriver/World.java
@@ -1,7 +1,0 @@
-package stepDefinitions.WebDriver;
-
-import java.util.HashMap;
-
-public class World {
-    public HashMap<String,Object> context = new HashMap<String, Object>();
-}
