--- src/test/java/utils/Location.java
@@ -1,19 +1,0 @@
-package utils;
-
-public class Location {
-private double lat;
-private double lng;
-public double getLat() {
-	return lat;
-}
-public void setLat(double lat) {
-	this.lat = lat;
-}
-public double getLng() {
-	return lng;
-}
-public void setLng(double lng) {
-	this.lng = lng;
-}
-
-}
