+++ testing.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE suite SYSTEM "testng.org/testng-1.0.dtd">
+
+<suite verbose="0" name="Default suite">
+    <test verbose="2" name="Default test" preserve-order="false">
+
+    </test>
+</suite>
